Refactored User Routes & Added Functions Folder
