+++ README.md
@@ -1,1 +1,24 @@
 # moodify
+
+## Server Development Rules
+ - token is always a parameter
+ - core mood is abbreviated `cm`
+ - associated moods are abbreviated `am1, am2, ..., am5`
+ - always use lowercase in the URL
+
+## Server Endpoints & Descriptions
+### User
+
+### Song
+
+### Playlist
+
+## Other Helpful Info
+### Status Codes
+ - 1xx informational response – the request was received, continuing process
+ - 2xx successful – the request was successfully received, understood, and accepted
+ - 3xx redirection – further action needs to be taken in order to complete the request
+ - 4xx client error – the request contains bad syntax or cannot be fulfilled
+ - 5xx server error – the server failed to fulfil an apparently valid request
+
+[For more info](https://en.wikipedia.org/wiki/List_of_HTTP_status_codes)
