+++ moodify-server/routes/user.js
@@ -1,102 +1,61 @@
+/*
+** Current to do on back-end is throwing errors in the try-catches
+** based on the return value of the functions.
+*/
+// Server Imports
 const express = require('express');
-const axios = require('axios');
-
 const router = express.Router();
-
-const spotify_url = 'https://api.spotify.com/v1';
-
 const User = require('../models/user');
-const Mood = require('../models/mood');
-const Angry = require("../models/angry")
-const Bad = require("../models/bad")
-const Content = require("../models/content")
-const Excited = require("../models/excited")
-const Happy = require("../models/happy")
-const Sad = require("../models/sad")
-const Song = require("../models/song")
+
+// Function Imports
+let { checkPlaylistFollow } = require('../functions/spotifyPlaylist');
+let { getUserId, getPlaylistFollow, userTop, getUserProfile } = require('../functions/spotifyUser');
+let { audioFeatures, idsToTracks, spotifyRecommend, filterTracks } = require('../functions/spotifySong');
+let { getSongByMood } = require('../functions/mongoSong');
+let { postUser, findUser, checkUser, logout, saveUser } = require('../functions/mongoUser');
+let { checkMood } = require('../functions/mongoMood');
 
 // http://localhost:5000/user/{token}
 // Get User Profile from Spotify
 // Return Raw User Profile from Spotify
-router.get('/:token', (req, res) => {
+router.get('/:token', async (req, res) => {
   console.log('running user profile api')
   const { token } = req.params;
-
-  axios.get(spotify_url + '/me', {
-    headers: { Authorization: `Bearer ${token}` },
-  })
-    .then((data) => {
-      console.log('successful user profile api')
-      console.log(token);
-      res.json(data.data);
-    })
-    .catch((err) => {
-      console.log('unsuccessful user profile api')
-      console.log(err);
-    });
+  try {
+    res.status(200).send(await getUserProfile(token));
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
+  }
 });
 
 // https://localhost:5000/user/tracks/:token
 // Get User's Top Tracks From Spotify
 // Returns User's Top Tracks in form of DB Song
-router.get('/tracks/:token', (req, res) => {
+router.get('/tracks/:token', async (req, res) => {
   console.log('running user top tracks');
   const { token } = req.params;
 
-  axios.get(spotify_url + '/me/top/tracks', {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got user top tracks!');
-    var items = data.data.items;
-    var toReturn = [];
-    items.forEach(element => {
-      toReturn.push({
-        "songID": element.id,
-        "songName": element.name,
-        "songArtist": element.artists,
-        "songAlbum": element.album.name,
-        "moodTag": "",
-        "popularity": element.popularity,
-        "performedBy": element.artists,
-        "writtenBy": element.artists,
-        "producedBy": element.artists
-      });
-    });
-    res.json(toReturn);
-  }).catch((err) => {
-    console.log('unsuccessful user top tracks')
-    console.log(err)
-  })
+  try {
+    res.status(200).send(await userTop(token, 'tracks'));
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
+  }
 })
 
 // https://localhost:5000/user/artists/:token
 // Get User's Top Artists From Spotify
 // Returns User's Top Artists by Name
-router.get('/artists/:token', (req, res) => {
+router.get('/artists/:token', async (req, res) => {
   console.log('running user top artists');
   const { token } = req.params;
-
-  axios.get(spotify_url + '/me/top/artists', {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got user top artists!');
-    var items = data.data.items;
-    var toReturn = [];
-    items.forEach(element => {
-      // console.log(element)
-      toReturn.push({
-        "id": element.id,
-        "name": element.name,
-        "genres": element.genres,
-        "popularity": element.popularity,
-        "followers": element.followers
-      });
-    });
-    res.status(200).json(toReturn)
-  }).catch((err) => {
-    console.log('unsuccessful user top artists')
-    console.log(err)
-  })
+  try {
+    res.status(200).send(await userTop(token, 'artists'));
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
+  }
 })
 
 // http://localhost:5000/user/dbprofile/{token}
@@ -106,18 +65,13 @@ router.get('/dbprofile/:token', async (req, res) => {
   console.log('running user info database retrieval');
   const { token } = req.params;
 
-  axios.get(spotify_url + '/me', {
-    headers: { Authorization: `Bearer ${token}` }
-  }).then((data) => {
-    console.log('Got user profile from spotify');
-    console.log(data.data.id);
-    return findUser(data.data.id);
-  }).then((data) => {
-    console.log(data);
-    res.json(data);
-  }).catch((err) => {
-    console.log(err);
-  })
+  try {
+    let id = await getUserId(token);
+    res.status(200).send(await findUser(id));
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
+  }
 })
 
 // http://localhost:5000/user/update/mood?type={core emotion}&token={token}
@@ -127,15 +81,13 @@ router.post('/update/mood', async (req, res) => {
   console.log('running user update mood info');
   const { type, token } = req.query;
 
-  axios.get(spotify_url + '/me', {
-    headers: { Authorization: `Bearer ${token}`}
-  }).then((data) => {
-    return checkMood(type, data.data.id);
-  }).then(() => {
-    res.sendStatus(200);
-  }).catch((err) => {
-    console.log(err);
-  })
+  try {
+    let id = await getUserId(token);
+    res.status(200).send(await checkMood(type, id));
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
+  }
 })
 
 //new user
@@ -148,16 +100,15 @@ router.post('/post', async (req, res) => {
     "userID": id
   };
 
-  if (await checkUser(user)) {
-    await postUser(user);
-    res.json({
-      "user was inserted": "into the db"
-    });
-  } else {
-    res.json({
-      "user was not inserted": "becuase they already exist"
-    });
-    console.log(user);
+  try {
+    if (await checkUser(user)) {
+      await postUser(user); // TODO: Figure out how to error throw this function
+      res.status(200).send("User Successfully Added");
+    } else
+      res.status(200).send("User Already Exists");
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
   }
 })
 
@@ -165,18 +116,10 @@ router.post('/logout', async (req, res) => {
   const { id } = req.query;
 
   try {
-    return User.findOneAndUpdate(
-      { "userID": id },
-      { "loggedin": false }
-    ).then((data) => {
-      if (data) {
-        res.sendStatus(200);
-      } else {
-        res.sendStatus(404);
-      }
-    })
+    res.status(200).send(await logout(id));
   } catch (err) {
     console.log(err);
+    res.sendStatus(400);
   }
 })
 
@@ -193,30 +136,48 @@ router.put('/recommended', async (req, res) =>  {
   };
 
   try {
-    return await User.findOne(
-      { "userID": user.userID }
-    ).then((data) => {
-      if (data) {
-        console.log("hereeeeee")
+    let data = await findUser(user.userID);
+    if (data) {
+      console.log(data.recommendedSongIDs);
+      if (!data.recommendedSongIDs.includes(user.songID)) {
+        data.recommendedSongIDs.push(user.songID);
+        data.numRecommendations++;
+        data = await saveUser(data);
         console.log(data)
-        if (!data.recommendedSongIDs.includes(user.songID)) {
-          data.recommendedSongIDs.push(user.songID);
-          data.numRecommendations++;
-          console.log(data)
-          data.save()
-          console.log("inserting new recommended song")
-          res.sendStatus(200)
-        } else {
-          console.log("user already has this recommended song")
-          //res.sendStatus(400)
-        }
+        console.log("inserting new recommended song")
+        res.sendStatus(200)
+      } else {
+        console.log("user already has this recommended song")
+        res.sendStatus(400)
       }
-    }).then(() => {
-      // res.sendStatus(200)
-    })
-  } catch (err) {
-    console.log(err);
+    }
+  } catch (error) {
+    console.log(error);
+    res.sendStatus(400);
   }
+})
+
+// http://localhost:5000/user/plalists/{token}?playlist={playlist_id}
+router.get('/playlists/:token', async (req, res) => {
+  const { token } = req.params;
+  const { playlist } = req.query;
+
+  if (playlist) {
+    try {
+      let id = await getUserId(token);
+      res.status(200).send(await checkPlaylistFollow(playlist, id, token));
+    } catch (err) {
+      console.log(err);
+      res.sendStatus(400);
+    }
+  } else {
+    let toReturn = await getPlaylistFollow(token);
+    if (toReturn)
+      res.status(200).send(toReturn);
+    else
+      res.sendStatus(400);
+  }
+
 })
 
 // http://localhost:5000/user/personal/{token}?cm={coremood}&am1={associatedmood}&am2={associatedmood}
@@ -278,260 +239,5 @@ router.get('/personal/:token', async (req, res) => {
     res.status(400);
   }
 })
-
-// Andrew Function's ATM, will be combined with the others soon
-
-// Request to spotify for user's top tracks
-// Returns list of ids of the user's top tracks
-async function userTopTracks(token) {
-  console.log('running user top tracks');
-
-  return await axios.get(spotify_url + '/me/top/tracks', {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got user top tracks!');
-    var items = data.data.items;
-    var toReturn = [];
-    items.forEach(element => {
-      toReturn.push({
-        "songID": element.id,
-        "songName": element.name,
-        "songArtist": element.artists,
-        "songAlbum": element.album.name,
-        "moodTag": "",
-      });
-    });
-    return toReturn;
-  }).catch((err) => {
-    console.log('unsuccessful user top tracks')
-    console.log(err)
-  })
-}
-
-// Request to mongo for song by mood and associated feels
-// Returns list of song ids
-async function getSongByMood(coreMood, associatedMood1, associatedMood2) {
-  coreMood.toLowerCase();
-  console.log("getting songs from " + coreMood);
-
-  switch (coreMood) {
-    case 'angry':
-      return await Angry.find();
-    case 'bad':
-      return await Bad.find();
-    case 'content':
-      return await Content.find();
-    case 'excited':
-      return await Excited.find();
-    case 'happy':
-      return await Happy.find();
-    case 'sad':
-      return await Sad.find();
-  }
-}
-
-// Request to spotify for audio features of list
-// Returns list of audio features by song id
-async function audioFeatures(list, token) {
-  let str = list.join()
-  return await axios.get(spotify_url + '/audio-features?ids=' + str, {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got audio features!');
-    return data.data;
-  }).catch((err) => {
-    console.log('Unsuccessful audio features!')
-    console.log(err)
-  })
-}
-
-// Uses features to filter 'toFilter' using 'list' features
-// Returns list of song ids that were filtered
-async function filterTracks(toFilter, list, features) {
-  // console.log(list)
-  list = features.audio_features.filter((item) => {
-    if (item) return list.includes(item.id);
-  })
-  // console.log(list)
-
-  // console.log(toFilter)
-  toFilter = features.audio_features.filter((item) => {
-    if (item) return toFilter.includes(item.id);
-  })
-
-  values = list.reduce((prev, curr) => {
-    return {
-      "tempo": (curr.tempo + prev.tempo) / 2,
-      "energy": (curr.energy + prev.energy) / 2,
-      "loudness": (curr.loudness + prev.loudness) / 2,
-      "danceability": (curr.danceability + prev.danceability) / 2,
-      "valence": (curr.valence + prev.valence) / 2,
-      "liveness": (curr.liveness + prev.liveness) / 2,
-    }
-  })
-
-  toFilter = toFilter.filter((item) => {
-    dance = Math.abs(item.danceability - values.danceability) < 0.1;
-    energy = Math.abs(item.energy - values.energy) < 0.5;
-    loud = Math.abs(item.loudness - values.loudness) < 10;
-    live = Math.abs(item.liveness - values.liveness) < 0.2;
-    val = Math.abs(item.valence - values.valence) < 0.2;
-    tempo = Math.abs(item.tempo - values.tempo) < 10;
-    // console.log(item.id, tempo, energy, loud, dance, val, live);
-    return tempo && energy && loud && dance && val && live;
-  })
-
-  if (toFilter.length == 0) {
-    toFilter.push(list[0])
-  }
-
-  return toFilter.map(item => item.id)
-}
-
-// Request to spotify for new songs
-// Returns json list of songs
-async function spotifyRecommend(combinedTracks, token) {
-  console.log('running spotify recommend');
-
-  str = combinedTracks.join();
-
-  return await axios.get(spotify_url + '/recommendations?seed_tracks=' + str, {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got recommendations!');
-    var items = data.data.tracks;
-    var toReturn = [];
-    items.forEach(element => {
-      toReturn.push(element.id);
-    });
-    return toReturn;
-  }).catch((err) => {
-    console.log('unsuccessful recommendations')
-    console.log(err)
-  })
-}
-
-// Request to spotify for new songs
-// Returns json list of songs
-async function idsToTracks(combinedTracks, token) {
-  console.log('running get tracks');
-
-  str = combinedTracks.join();
-
-  return await axios.get(spotify_url + '/tracks?ids=' + str, {
-    headers: { Authorization: `Bearer ${token}` },
-  }).then((data) => {
-    console.log('Got tracks!');
-    var items = data.data.tracks;
-    var toReturn = [];
-    items.forEach(element => {
-      toReturn.push({
-        "songID": element.id,
-        "songName": element.name,
-        "songArtist": element.artists,
-        "songAlbum": element.album.name,
-        "moodTag": "",
-      });
-    });
-    return toReturn;
-  }).catch((err) => {
-    console.log('unsuccessful get tracks')
-    console.log(err)
-  })
-}
-
-//FUNCTIONS
-
-async function postUser(user) {
-  try {
-    await new User(
-      {
-        "userID": user.userID,
-        "logins": 1,
-        "numRecommendations": 0,
-        "loggedin": true
-      }
-    ).save();
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-
-async function checkUser(user) {
-
-  try {
-    return await User.findOne(
-      { "userID": user.userID }
-    ).then((data) => {
-      if (data) {
-        // console.log(data)
-        data.logins += 1;
-        data.loggedin = true;
-        data.save()
-        return false;
-      } else {
-        return true;
-      }
-    })
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-async function findUser(userID) {
-  try {
-    let promise_obj = await User.findOne(
-      { "userID": userID }//,
-      //'userID logins',
-    )
-    console.log(promise_obj);
-    return {
-      "userID": promise_obj.userID,
-      "logins": promise_obj.logins,
-      "recommendedSongsIDs": promise_obj.recommendedSongIDs,
-      "numRecommendations": promise_obj.numRecommendations
-    };
-  } catch (err) {
-    console.log(err);
-  }
-}
-
-// checks if mood is already in db and updates, if not it'll create it.
-async function checkMood(mood, userID) {
-  try {
-    let promise_obj = await Mood.findOne(
-      { "userID": userID, "type": mood }
-    )
-    if (promise_obj) {
-      let arr = promise_obj.timeStamp;
-      let timeNow = new Date();
-      arr.push(timeNow.getTime());
-      await Mood.updateOne(
-        {
-          "userID": userID,
-          "type": mood
-        },
-        {
-          "timeStamp": arr,
-          "totalCount": promise_obj.totalCount + 1
-        }
-      )
-    } else {
-      let timeNow = new Date();
-      let userMood = new Mood(
-        {
-          "userID": userID,
-          "type": mood,
-          "timeStamp": [timeNow.getTime()],
-          "totalCount": 1,
-        }
-      )
-      await userMood.save()
-    }
-  } catch (err) {
-    console.log(err)
-  }
-}
 
 module.exports = router;
